+++ custom.d.ts
@@ -1,0 +1,5 @@
+declare module "*.svg" {
+  import React from "react"
+  const SVG: React.VFC<React.SVGProps<SVGSVGElement>>
+  export default SVG
+}
