+++ src/assets/icons.tsx
@@ -1,0 +1,406 @@
+import React from "react"
+
+export const Bienestar = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="81"
+    height="93"
+    viewBox="0 0 81 93"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M43.6101 61.0844C39.9201 61.6844 36.3601 61.5144 32.9701 59.8544C27.5001 57.1644 24.5101 51.2444 25.6501 45.4644C26.8801 39.2244 31.9801 34.7544 38.2901 34.3844C38.9101 34.3444 39.5301 34.3844 40.1501 34.3844C45.9901 34.4144 50.7601 36.4944 53.4801 41.9344C57.5301 50.0144 52.5001 59.6344 43.6101 61.0844Z"
+      fill={color}
+    />
+    <path
+      d="M42.0201 27.9945C40.0201 28.3145 38.0901 28.2245 36.2501 27.3245C33.2801 25.8645 31.6701 22.6545 32.2801 19.5245C32.9501 16.1445 35.7101 13.7245 39.1301 13.5245C39.4701 13.5045 39.8001 13.5245 40.1401 13.5245C43.3101 13.5345 45.8901 14.6645 47.3601 17.6145C49.5601 21.9945 46.8301 27.2045 42.0101 27.9945H42.0201Z"
+      fill={color}
+    />
+    <path
+      d="M41.0901 7.23444C40.0901 7.39444 39.1201 7.35444 38.2101 6.90444C36.7301 6.17444 35.9201 4.57444 36.2301 3.00444C36.5601 1.31444 37.9501 0.104444 39.6601 0.00444444C39.8301 -0.00555556 40.0001 0.00444444 40.1701 0.00444444C41.7501 0.00444444 43.0401 0.574444 43.7801 2.05444C44.8801 4.24444 43.5201 6.85444 41.1101 7.24444L41.0901 7.23444Z"
+      fill={color}
+    />
+    <path
+      d="M79.2301 65.2245C79.8801 65.2245 80.4001 65.8245 80.2901 66.4645C79.7701 69.4845 78.7801 72.2245 77.3301 74.6345C75.5801 77.6345 73.3501 80.2845 70.6501 82.5345C69.8701 83.1745 69.0601 83.7545 68.2301 84.3445C64.4501 86.9645 60.1301 88.8545 55.2701 90.1345C50.4001 91.4145 45.3701 92.0245 40.1601 92.0245C34.9501 92.0245 29.9501 91.4145 25.0501 90.1345C20.1501 88.8545 15.8101 86.9645 12.0601 84.3445C11.2301 83.7645 10.4501 83.1745 9.70012 82.5345C7.03012 80.2845 4.80012 77.6345 3.02012 74.6345C1.77012 72.5445 0.880121 70.2345 0.320121 67.6845C0.290121 67.5445 0.200122 67.1645 0.190122 67.0245C0.130122 66.5145 -0.249878 65.2245 0.270122 65.2345C14.1501 65.3545 79.2301 65.2345 79.2301 65.2345V65.2245Z"
+      fill={color}
+    />
+  </svg>
+)
+
+export const Cuerpo = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="91"
+    height="90"
+    viewBox="0 0 91 90"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M90.1244 75.0789C90.3244 76.2889 89.4844 77.4089 88.4944 77.1489C83.8744 75.9489 79.3344 76.5089 75.2444 77.9289C70.6244 79.5789 66.4744 82.2689 62.1644 84.6489C61.8544 84.8189 61.4944 85.0089 61.1844 85.2089C56.5344 87.6989 51.4944 89.6889 45.8644 89.3589C45.6144 89.3289 45.3344 89.3289 45.0844 89.2989C44.8044 89.3289 44.5544 89.3289 44.2744 89.3589C38.6744 89.6989 33.6244 87.7089 28.9744 85.2089C28.6344 85.0089 28.3044 84.8189 27.9944 84.6489C23.6844 82.2689 19.5344 79.5789 14.9144 77.9289C10.8244 76.4989 6.25442 75.9389 1.63442 77.1489C0.684421 77.3989 -0.185579 76.2789 0.0344213 75.0789C0.454421 72.6989 1.12442 70.5089 2.05442 68.5489C3.73442 64.9089 6.00442 61.8489 8.80442 59.4189C11.2144 57.3189 13.9844 55.6889 17.0444 54.4889C16.8444 54.3189 16.6544 54.1789 16.4844 54.0089C13.6544 51.5689 11.3844 48.5189 9.70442 44.9089C8.19442 41.6889 7.32442 37.9089 7.18442 33.6189C7.15442 33.1189 7.32442 32.6089 7.60442 32.2489C11.8944 27.0989 17.7744 24.4089 23.0144 24.0989C28.6444 23.7389 33.6644 25.7489 38.3344 28.2489C38.6444 28.4189 38.9744 28.6089 39.3144 28.7789C41.2744 29.8689 43.2344 31.0489 45.1944 32.1389C47.1844 31.0489 49.1144 29.8689 51.1044 28.7789C51.4444 28.6089 51.7744 28.4189 52.0844 28.2489C56.7644 25.7589 61.7744 23.7389 67.4044 24.0989C72.6444 24.4089 78.4944 27.0989 82.8144 32.2489C83.0944 32.6089 83.2644 33.1189 83.2344 33.6189C83.0944 37.9089 82.2244 41.6889 80.7144 44.9089C79.0344 48.5189 76.7644 51.5789 73.9344 54.0089C73.7144 54.2089 73.4844 54.3689 73.2644 54.5389C76.2944 55.7489 78.9544 57.3389 81.3644 59.4189C84.1644 61.8589 86.4044 64.9089 88.1144 68.5489C89.0344 70.5089 89.7144 72.6989 90.1344 75.0789H90.1244Z"
+      fill={color}
+    />
+    <path
+      d="M46.9544 14.5789C44.9444 14.9089 42.9944 14.8089 41.1444 13.9089C38.1544 12.4389 36.5244 9.20889 37.1444 6.05889C37.8244 2.64889 40.6144 0.208889 44.0544 0.00888889C44.3944 -0.0111111 44.7344 0.00888889 45.0744 0.00888889C48.2644 0.0188889 50.8644 1.15889 52.3444 4.12889C54.5544 8.53889 51.8144 13.7889 46.9544 14.5789Z"
+      fill={color}
+    />
+  </svg>
+)
+
+export const EmailIcon = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_1"
+    data-name="Layer 1"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 500 500"
+    {...props}
+  >
+    <defs></defs>
+    <path
+      fill={color}
+      d="M58.34,363.04v-175.91c.56.24,1.17.4,1.66.73,51.86,34.57,103.71,69.16,155.56,103.74,22.6,15.07,46.03,15.19,68.52.2,51.25-34.13,102.47-68.3,153.7-102.45,1.09-.73,2.22-1.41,3.8-2.42v3.05c0,55.88.02,111.77-.05,167.65,0,3.08-.37,6.25-1.12,9.24-4.14,16.43-18.04,26.7-36.06,26.71-51.02.03-102.04,0-153.05,0s-103.78-.02-155.67.02c-10.16,0-19.1-2.98-26.56-10.05-5.94-5.62-8.89-12.77-10.73-20.52Z"
+    />
+    <path
+      fill={color}
+      d="M58.34,136.98c.62-2.14,1.15-4.3,1.86-6.41,4.72-14.02,17.23-23.3,32.44-24.09,1.24-.07,2.49-.05,3.74-.05,102.41,0,204.83.02,307.24-.04,10.86,0,20.43,2.96,28.05,10.93,15.2,15.9,12.77,40.3-5.48,52.68-23.21,15.74-46.64,31.17-69.97,46.73-27.9,18.6-55.79,37.21-83.7,55.8-16.05,10.69-29.03,10.69-45.06,0-49.17-32.77-98.24-65.69-147.57-98.21-10.52-6.94-18.97-14.82-21.56-27.62,0-3.24,0-6.49,0-9.73Z"
+    />
+  </svg>
+)
+
+export const Espiritu = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="89"
+    height="97"
+    viewBox="0 0 89 97"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M87.5135 36.9913C82.7335 7.85132 51.2535 -8.61868 24.7735 4.65132C6.96346 13.5713 0.113461 29.1613 0.0434609 48.3013C0.0434609 50.3513 -0.0665386 52.3713 0.0634614 54.4013C1.28346 75.0813 15.9335 91.7913 36.3435 95.8213C55.2635 99.5413 74.6735 89.7713 83.4635 71.8313C88.9135 60.7413 89.4535 49.0613 87.5135 36.9813V36.9913ZM41.3935 28.0113C41.6335 26.7813 42.6435 25.8913 43.9035 25.8113C44.0335 25.7913 44.1535 25.8113 44.2635 25.8113C45.4235 25.8113 46.3735 26.2313 46.9135 27.3013C47.7335 28.9213 46.7335 30.8213 44.9535 31.1113C44.2235 31.2213 43.5235 31.2013 42.8435 30.8613C41.7535 30.3313 41.1535 29.1513 41.3935 28.0113ZM38.5035 40.1213C38.9935 37.6513 41.0235 35.8713 43.5335 35.7313C43.7735 35.7113 44.0235 35.7313 44.2635 35.7313C46.5835 35.7313 48.4735 36.5713 49.5635 38.7313C51.1835 41.9413 49.1835 45.7613 45.6435 46.3413C44.1735 46.5813 42.7535 46.5013 41.4135 45.8513C39.2535 44.7813 38.0535 42.4413 38.5035 40.1313V40.1213ZM46.8035 70.5913C44.0935 71.0313 41.4835 70.9013 38.9935 69.6813C34.9835 67.7013 32.8035 63.3613 33.6335 59.1313C34.5435 54.5513 38.2635 51.2913 42.8935 51.0113C43.3435 50.9713 43.8035 51.0113 44.2535 51.0113C48.5435 51.0313 52.0435 52.5513 54.0235 56.5513C57.0035 62.4713 53.3135 69.5213 46.7935 70.5913H46.8035Z"
+      fill={color}
+    />
+  </svg>
+)
+
+export const InstagramIcon = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_1"
+    data-name="Layer 1"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 500 500"
+    {...props}
+  >
+    <defs></defs>
+    <path
+      fill={color}
+      d="M338.57,414.3h-177.09c-6.15-1.3-12.47-2.06-18.41-3.99-34.55-11.23-57.16-42.24-57.26-78.7-.15-54.43-.08-108.87-.02-163.3.02-17.48,4.86-33.59,15.21-47.75,17.02-23.29,40.18-34.8,68.98-34.83,52.93-.05,105.85-.05,158.78.04,5.54,0,11.16.25,16.59,1.24,30.35,5.53,51.54,22.76,63.11,51.27,2.97,7.31,3.94,15.42,5.83,23.16v177.13c-.41,2.29-.79,4.59-1.22,6.88-5.68,30.16-22.76,51.28-51.03,62.91-7.38,3.04-15.62,4.01-23.46,5.94ZM181.57,250.02c.02,37.61,30.49,68.14,68.31,68.46,37.43.31,68.64-30.84,68.61-68.48-.03-37.61-30.5-68.15-68.31-68.46-37.46-.3-68.62,30.8-68.6,68.48ZM339.01,181.46c11.22,0,20.34-9.02,20.51-20.29.17-11.16-9.27-20.68-20.53-20.68-11.16,0-20.35,9.16-20.43,20.37-.09,11.4,9.05,20.6,20.45,20.6Z"
+    />
+  </svg>
+)
+
+export const Isotipo = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_2"
+    data-name="Layer 2"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 49.51 48.72"
+    {...props}
+  >
+    <g id="Layer_1-2" data-name="Layer 1">
+      <g>
+        <path
+          fill={color}
+          d="M27.32,47.2c-.48-1.32-.29-2.74-.28-4.13,0,.69,0,1.39,0,2.08,0,.68.14,1.35.28,2.05Z"
+        />
+        <path
+          fill={color}
+          d="M49.51,48.61c-1.05,0-2.13-.12-3.17.02-3.03.43-5.23-.75-7.36-2.89-5.75-5.78-7.59-16.13-3.67-23.23,1.63-2.94,3.8-5.56,5.41-8.5.85-1.57,1.34-3.49,1.45-5.28.19-3.29-1.9-5.81-5.21-6.86-3.04-.97-6.17-.98-9.43-.87-.14,0-.29,0-.44.01-.03.52-.08,1-.08,1.46,0,13.45-.01,26.92.02,40.37v.23c-.01,1.39-.2,2.81.28,4.13,0,0,.01.01,0,.02.09.42.18.85.25,1.28h-5.64c.08-.46.18-.9.27-1.35.47-1.3.29-2.71.27-4.08,0-.08,0-.16,0-.24.04-13.45.02-26.92.01-40.37,0-.47-.04-.94-.07-1.46-.14,0-.29-.01-.44-.01-3.26-.11-6.39-.09-9.43.87-3.31,1.05-5.39,3.57-5.2,6.86.11,1.79.59,3.71,1.45,5.28,1.6,2.94,3.79,5.57,5.41,8.5,3.92,7.1,2.08,17.45-3.67,23.23-2.14,2.14-4.33,3.32-7.37,2.89-1.04-.14-2.11-.02-3.17-.02.7-.75,1.41-1.2,2.1-1.68,4.2-2.93,8.24-6.05,11.02-10.44,2.69-4.27,3.21-9.67-.7-13.84-1.45-1.55-2.94-3.07-4.38-4.63-1.33-1.43-2.73-2.81-3.87-4.38-2.18-3.01-2.06-7.29.14-9.73C6.12,1.91,8.57.9,11.13.69c3.57-.28,7.15-.43,10.67-.57,1-.04,1.98-.08,2.96-.12.97.04,1.96.08,2.95.12,3.52.14,7.11.29,10.68.57,2.56.21,5.01,1.22,6.82,3.22,2.2,2.44,2.31,6.73.14,9.73-1.13,1.57-2.54,2.94-3.86,4.38-1.44,1.56-2.92,3.09-4.38,4.63-3.91,4.17-3.4,9.57-.7,13.84,2.77,4.39,6.82,7.51,11.01,10.44.69.48,1.4.93,2.1,1.68Z"
+        />
+        <path
+          fill={color}
+          d="M22.48,43.08c.01,1.37.19,2.77-.27,4.08.14-.68.27-1.34.27-2.01,0-.69,0-1.38,0-2.07Z"
+        />
+      </g>
+    </g>
+  </svg>
+)
+
+export const LocationIcon = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_1"
+    data-name="Layer 1"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 500 500"
+    {...props}
+  >
+    <path
+      fill={color}
+      d="M247.75,442c-22.16-21.13-42.45-43.95-61.5-67.87-22.29-27.97-42.48-57.32-58.11-89.59-9.88-20.39-17.77-41.45-20.78-64.06-4.43-33.28,2.24-64.39,19.87-92.79,23.27-37.49,57.16-60.01,100.75-67.2,57.61-9.5,116.67,18.62,145.99,69.08,23.07,39.71,25.55,81.25,10.63,124.27-11.13,32.09-28.01,61.11-47.33,88.81-24.06,34.49-51.41,66.22-81.25,95.84-1.22,1.21-2.51,2.34-3.77,3.51h-4.5ZM250.13,268.33c35.98.05,65.74-29.48,65.86-65.35.13-36.52-29.36-66.38-65.56-66.38-36.84,0-66.34,29.27-66.42,65.9-.08,36.24,29.59,65.78,66.12,65.83Z"
+    />
+  </svg>
+)
+
+export const Logo = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_2"
+    data-name="Layer 2"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 354.67 74.4"
+    {...props}
+  >
+    <g id="Layer_1-2" data-name="Layer 1">
+      <g>
+        <g>
+          <path
+            fill={color}
+            d="M190.43,19.3c2.26-.61,4.99-1.33,7.9-2.11v8.72c2.21-1.93,4.14-3.9,6.35-5.49,5.48-3.96,13.14-4.38,18.21-1.18,2.31,1.46,3.14,2.54,5.36,7.01,1.62-1.52,3.24-3.11,4.94-4.62,5.16-4.59,12.89-5.81,18.75-2.93,4.81,2.36,6.81,6.64,7.06,11.62.36,7.01.3,14.05.36,21.08.05,5.77-.05,11.55.05,17.32.03,1.58.56,3.15.88,4.83h-8.76c.33-1.69.96-3.43.97-5.17.07-11.88.04-23.75,0-35.63-.05-11.67-8.64-13.86-16.47-10.63-2.52,1.04-4.54,3.3-6.71,5.1-.41.34-.61,1.16-.6,1.75.09,6.42.29,12.85.36,19.27.07,6.57-.06,13.14.06,19.7.03,1.87.63,3.73.98,5.65h-8.85c.36-1.41.99-2.78,1-4.15.05-12.6.11-25.21-.07-37.81-.15-10.48-8.56-12.39-15.71-9.82-2.95,1.06-5.3,2.99-7.25,5.38-.49.6-.82,1.53-.82,2.31-.05,12.94-.07,25.88.01,38.81.01,1.74.65,3.47,1.01,5.27h-8.69c.29-1.84.81-3.67.82-5.5.07-14.6.07-29.19-.01-43.79,0-1.53-.68-3.06-1.14-4.97Z"
+          />
+          <path
+            fill={color}
+            d="M.48.76c8.09.37,16.51.56,24.9,1.22,3.84.3,7.49,1.82,10.2,4.82,3.3,3.65,3.47,10.06.21,14.57-1.7,2.35-3.81,4.4-5.78,6.55-2.16,2.34-4.38,4.62-6.56,6.94-5.85,6.23-5.08,14.32-1.05,20.7,4.15,6.58,10.21,11.23,16.48,15.62,1.03.72,2.1,1.4,3.14,2.52-1.58,0-3.19-.18-4.75.03-4.54.64-7.83-1.12-11.02-4.34-8.61-8.65-11.36-24.13-5.49-34.76,2.43-4.39,5.69-8.32,8.1-12.72,1.28-2.35,2.01-5.22,2.16-7.9.29-4.92-2.83-8.69-7.79-10.27-4.76-1.52-9.65-1.47-14.76-1.29-.04.78-.12,1.49-.12,2.19,0,21.29-.02,42.58.03,63.87,0,1.64.51,3.28.81,5.01H0c2.17-2.55,1.56-5.58,1.57-8.48.03-19.43.04-38.87-.03-58.3,0-1.81-.62-3.63-1.06-5.99Z"
+          />
+          <path
+            fill={color}
+            d="M100.77,44.58c.32-10.67,5.37-19.82,16.32-25.2,10.96-5.39,25.93-.93,32.4,9.6,7,11.41,6.99,22.99-.35,34.18-4.4,6.71-10.97,10.42-19.09,11.01-11.29.83-19.74-3.83-25.52-13.45-2.42-4.02-3.59-8.83-3.76-16.15ZM108.11,45.84c-.05,4.69.71,9.23,2.5,13.58,2.18,5.3,5.37,9.84,10.77,12.18,7.5,3.25,15.39.53,20.27-6.66,7.3-10.75,7.24-27.82-.11-38.53-6.89-10.03-18.41-10.87-26.41-1.76-5.29,6.02-6.96,13.38-7.02,21.19Z"
+          />
+          <path
+            fill={color}
+            d="M351.65,71.78c-6.03-3.41-11.85-7.02-15.84-12.13-3.88-4.95-4.61-11.24,1.01-16.08,2.09-1.8,4.23-3.57,6.3-5.38,1.89-1.67,3.92-3.26,5.55-5.09,3.13-3.5,2.97-8.48-.2-11.31-2.61-2.33-6.12-3.51-9.8-3.74-8.06-.52-16.16-.66-23.93-.95.42,1.83,1.01,3.24,1.01,4.65.04,8.81.05,17.63.05,26.44-.41.84-.84,1.75-1.29,2.69-1.32,2.78-2.78,5.78-4.1,7.76-2.34,3.53-5.08,5.95-5.08,5.95l-.52.53-.03.02-.02.02c-1.54,1.26-2.84,1.95-4.61,2.58-10.31,3.65-20.91-1.18-25.15-11.4-3.14-7.58-3.13-15.29-1.41-23.3h34.74c.58-5.77-2.52-11.62-7.24-13.8-10.42-4.82-23.2-1.23-29.82,9.38-4.53,7.26-5.65,15.22-4.19,23.51,1.96,11.11,8.75,19.54,19.52,21.79,5.3,1.11,10.35.29,14.43-2.2.55-.33,1.09-.7,1.6-1.1,0,0,.01,0,.02-.01.74-.57,1.43-1.2,2.07-1.89l.05-.05c.14-.14.38-.37.7-.67.01-.01.01-.02.02-.02,1.38-1.33,4.13-4.08,6.15-7.31.97-1.55,1.93-3.21,2.86-4.91.44-.8.87-1.61,1.29-2.41,0,4.56,0,9.12-.01,13.68,0,2.25.58,4.6-1.51,6.58h8.84c-.28-1.35-.77-2.62-.77-3.89-.05-16.54-.04-33.07-.03-49.6,0-.55.07-1.09.11-1.7,4.91-.15,9.61-.18,14.19,1,4.76,1.22,7.76,4.14,7.48,7.97-.15,2.08-.85,4.31-2.08,6.13-2.31,3.42-5.45,6.47-7.78,9.89-5.64,8.25-2.99,20.27,5.28,26.99,3.07,2.49,6.23,3.86,10.59,3.36,1.5-.16,3.04-.02,4.57-.02-1.01-.88-2.03-1.4-3.02-1.96ZM294.71,18.83c3.76.8,6.18,3.32,6.64,6.93.36,2.87-.82,4.74-3.56,5.57-.7.2-1.43.41-2.14.41-6.38.05-12.76.06-19.14.08h-2.33c2.6-9.51,11.34-14.93,20.53-12.99Z"
+          />
+          <path
+            fill={color}
+            d="M77.68,61.14c-.09.15-.19.3-.29.44.18-.21.36-.43.55-.66-.11.07-.2.13-.26.22Z"
+          />
+          <path
+            fill={color}
+            d="M96.72,6.01c4.41-.45,8.71.23,12.91,1.98.57-2.6-.36-4.44-2.22-5.88-3.61-2.79-9.12-2.76-13.29-.22-4.06,2.48-6.06,6.32-7.55,10.57-1.06,2.99-2.94,4.99-6.28,5.29-.45.04-.89.26-1.33.39.01.23.02.47.03.7,1.94.1,3.89.21,6.02.33v2.47c0,8.83.01,17.66.01,26.5-.42.85-.86,1.78-1.32,2.74-1.32,2.78-2.78,5.78-4.09,7.76-.55.83-1.12,1.6-1.67,2.28-.19.23-.37.45-.55.66-.01.02-.02.03-.04.04-1.07,1.62-2.38,2.94-3.84,3.99-5.17,3.73-12.18,3.99-17.7,2.12-5.59-1.89-9.23-5.86-11.48-11.1-3.28-7.63-3.44-15.5-1.47-23.59h20.09c4.39,0,8.78,0,13.17.01.98.01,1.66-.09,1.63-1.37-.11-5.93-3.42-11.07-8.97-13.08-11.95-4.32-24.8,1.14-30.27,13.74-3.94,9.08-3.64,18.29.11,27.41,4.4,10.73,16.16,16.72,27.32,13.92,2.06-.51,4.13-1.63,5.96-3.09.11-.06.23-.12.34-.18.01,0,.02-.01.03-.01.86-.71,1.66-1.5,2.39-2.39.01-.01.01-.02.02-.02,1.38-1.33,4.13-4.08,6.16-7.31.96-1.55,1.92-3.21,2.86-4.91.44-.81.88-1.63,1.31-2.44,0,4.49-.02,8.99-.04,13.49-.01,2.27-.55,4.54-.85,6.79h8.52c-.27-2.02-.77-3.99-.78-5.95-.06-15.39-.03-30.77-.03-46.16v-2.34c5.16-.4,10.1.22,14.94-.84-.89-.48-1.77-.44-2.65-.48-4.16-.22-8.37-.13-12.45-.8-3.37-.55-4.47-3.9-2.67-6.79,1.76-2.84,4.59-3.91,7.72-4.23ZM48.63,22.96c4.98-4.42,10.76-5.77,17.12-3.43,3.33,1.22,5.07,4.08,4.89,7.38-.11,2.07-1.18,3.46-3.09,4.18-.79.3-1.65.61-2.49.61-7.15.05-14.31.03-21.67.03.93-3.52,2.65-6.48,5.24-8.77Z"
+          />
+          <path
+            fill={color}
+            d="M185.05,17.39c-.38,1.97-.64,3.81-1.13,5.57-.14.51-.91,1.13-1.47,1.22-7.61,1.19-11.65,6.16-13.54,13.07-.86,3.14-1.15,6.51-1.23,9.78-.18,6.96-.12,13.92-.01,20.89.03,1.89.58,3.77.89,5.68h-8.56c.27-1.94.76-3.83.77-5.72.06-14.32.02-28.65.05-42.97,0-1.97.03-3.87-2.18-5.34,3.14-.86,5.85-1.6,8.84-2.41v18.99c4.06-14.04,10.02-18.78,17.56-18.77Z"
+          />
+        </g>
+        <g>
+          <path
+            fill={color}
+            d="M305.47,68c-.74.89-1.53,1.68-2.4,2.39,0,0-.01.01-.02.01-.13.07-.27.14-.4.21.74-.57,1.43-1.2,2.07-1.89l.05-.05c.14-.14.38-.37.7-.67Z"
+          />
+          <path
+            fill={color}
+            d="M77.35,61.62c-1.07,1.62-2.38,2.94-3.84,3.99l1.01-1.02s1.28-1.13,2.83-2.97Z"
+          />
+          <path
+            fill={color}
+            d="M304.77,68.67l-.05.05c-.64.69-1.33,1.32-2.07,1.89-.01.01-.02.01-.02.01-.51.4-1.05.77-1.6,1.1.15-.06.31-.12.47-.18.54-.34,1.05-.72,1.55-1.14.01,0,.02-.01.02-.01.87-.71,1.66-1.5,2.4-2.39-.32.3-.56.53-.7.67Z"
+          />
+        </g>
+      </g>
+    </g>
+  </svg>
+)
+
+export const LogoIzqFooter = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_2"
+    data-name="Layer 2"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 275.39 94.15"
+    {...props}
+  >
+    <g id="Layer_1-2" data-name="Layer 1">
+      <g>
+        <path
+          fill={color}
+          d="M13.07,93.84h-1.56l-3.16-9.94-3.06,9.94h-1.6l-3.69-12h1.68l2.84,9.94,3.04-9.94h1.56l3.16,9.94,2.8-9.94h1.65l-3.67,12Z"
+        />
+        <path
+          fill={color}
+          d="M20.09,93.84v-12h8.26v1.41h-6.73v3.71h5.85v1.39h-5.85v4.1h6.94v1.39h-8.47Z"
+        />
+        <path fill={color} d="M31.88,93.84v-12h1.6v10.57h5.81v1.43h-7.41Z" />
+        <path fill={color} d="M42.43,93.84v-12h1.6v10.57h5.81v1.43h-7.41Z" />
+        <path
+          fill={color}
+          d="M61.1,93.84l-6.58-9.76.15,9.76h-1.51v-12h1.51l6.49,9.62-.07-9.62h1.44v12h-1.44Z"
+        />
+        <path
+          fill={color}
+          d="M66.91,93.84v-12h8.26v1.41h-6.73v3.71h5.85v1.39h-5.85v4.1h6.94v1.39h-8.47Z"
+        />
+        <path
+          fill={color}
+          d="M87.04,92.86c-.66.65-1.75,1.29-3.79,1.29-.39,0-2.02-.02-3.28-.83-.37-.24-.71-.56-.97-.9-.71-.95-.8-2.04-.83-2.41l1.72-.22c.02.36.05,1.36.82,2.16.42.44,1.22.95,2.64.95.17,0,.63,0,1.1-.14.51-.14,1.89-.61,1.89-2.26,0-.8-.34-1.14-.61-1.36-.48-.37-1.11-.51-1.66-.63l-1.04-.22c-.87-.19-1.84-.39-2.64-.83-1.44-.78-1.63-1.87-1.63-2.58,0-.95.42-1.75,1.14-2.35,1.22-1.02,2.72-.99,3.21-.99,1.16,0,2.28.19,3.21.9.56.43.83.83.97,1.07.32.54.46,1.1.56,1.72l-1.61.1c-.02-.24-.05-.48-.12-.71-.24-.85-1-1.84-2.97-1.84-.76,0-1.8.17-2.33.8-.25.29-.41.7-.41,1.09,0,.17.03.54.29.88.46.63,1.39.82,2.09.99l1.33.32c.31.07.63.15.94.24.37.08,1.84.41,2.57,1.63.17.29.44.82.44,1.67,0,.48-.07,1.55-1,2.46Z"
+        />
+        <path
+          fill={color}
+          d="M100.08,92.86c-.66.65-1.75,1.29-3.79,1.29-.39,0-2.02-.02-3.28-.83-.37-.24-.71-.56-.97-.9-.71-.95-.8-2.04-.83-2.41l1.72-.22c.02.36.05,1.36.82,2.16.43.44,1.23.95,2.64.95.17,0,.63,0,1.11-.14.51-.14,1.89-.61,1.89-2.26,0-.8-.34-1.14-.61-1.36-.48-.37-1.11-.51-1.67-.63l-1.04-.22c-.87-.19-1.84-.39-2.64-.83-1.44-.78-1.63-1.87-1.63-2.58,0-.95.42-1.75,1.14-2.35,1.23-1.02,2.72-.99,3.21-.99,1.16,0,2.28.19,3.21.9.56.43.83.83.97,1.07.32.54.46,1.1.56,1.72l-1.61.1c-.02-.24-.05-.48-.12-.71-.24-.85-1-1.84-2.97-1.84-.77,0-1.8.17-2.33.8-.25.29-.41.7-.41,1.09,0,.17.03.54.29.88.46.63,1.39.82,2.09.99l1.33.32c.3.07.63.15.93.24.37.08,1.84.41,2.57,1.63.17.29.44.82.44,1.67,0,.48-.07,1.55-1,2.46Z"
+        />
+        <path
+          fill={color}
+          d="M122.25,90.44c-.31.78-1.34,2.74-3.71,3.43-.85.25-1.6.27-1.92.27-2.12,0-3.33-.85-3.98-1.44-.8-.75-1.89-2.24-1.89-4.86,0-3.74,2.26-6.31,6.07-6.31.68,0,2.94.05,4.42,1.78.49.58.8,1.19,1.05,1.89l-1.61.32c-.2-.58-.39-.9-.54-1.12-.42-.6-1.44-1.6-3.33-1.6-.29,0-1.02.03-1.75.34-2.52,1.07-2.53,3.94-2.53,4.61,0,2.89,1.38,4.25,2.55,4.74.31.12.88.31,1.75.31.2,0,.68,0,1.22-.15.48-.14,1.6-.6,2.26-1.84.19-.36.34-.75.49-1.53l1.73.2c-.03.17-.12.53-.29.95Z"
+        />
+        <path fill={color} d="M125.77,93.84v-12h1.6v10.57h5.81v1.43h-7.41Z" />
+        <path
+          fill={color}
+          d="M145.45,90.9c-.1.54-.48,1.7-1.75,2.52-.49.31-1.38.73-2.79.73-.63,0-1.34-.08-1.96-.32-1.09-.41-1.97-1.31-2.33-2.43-.17-.58-.17-1-.17-2.12v-7.43h1.6v7.21c0,.46-.02.93.05,1.39.03.27.17,1.07.78,1.65.29.27.63.46,1,.58.42.12.9.12,1.04.12.29,0,1.05,0,1.78-.48,1.19-.76,1.22-2.06,1.22-3.3v-7.17h1.61v7.67c0,.12.02.83-.1,1.39Z"
+        />
+        <path
+          fill={color}
+          d="M157.15,93.64c-.71.22-1.22.2-2.58.2h-4.69v-12h5.65c.97,0,1.82.2,2.58.87.66.56.99,1.33.99,2.21,0,.78-.25,1.46-.82,2.01-.25.24-.54.41-.65.46-.07.03-.19.08-.37.14.34.1.65.22.95.43.51.34,1.24,1.09,1.24,2.58-.02,2.16-1.53,2.89-2.29,3.11ZM156.66,83.56c-.48-.29-.85-.31-1.85-.31h-3.42v3.64h2.52c1,0,1.82-.02,2.19-.12.44-.1,1.31-.53,1.31-1.77,0-.32-.02-1-.75-1.44ZM157.64,89.54c-.14-.32-.53-.99-1.44-1.19-.53-.12-.88-.1-1.87-.1h-2.94v4.18h3.55c.66,0,1.21-.03,1.44-.1.1-.02.36-.08.61-.24.36-.22.8-.66.8-1.68,0-.15,0-.51-.15-.87Z"
+        />
+      </g>
+      <g>
+        <g>
+          <path
+            fill={color}
+            d="M147.89,14.98c1.76-.47,3.87-1.04,6.14-1.64v6.77c1.72-1.5,3.22-3.03,4.93-4.27,4.25-3.08,10.2-3.4,14.14-.92,1.79,1.13,2.44,1.98,4.16,5.44,1.26-1.18,2.52-2.42,3.84-3.59,4-3.57,10.01-4.51,14.56-2.27,3.73,1.83,5.28,5.15,5.48,9.02.28,5.44.23,10.91.28,16.36.04,4.48-.04,8.96.04,13.44.02,1.22.44,2.44.69,3.75h-6.8c.26-1.31.75-2.66.76-4.02.06-9.22.03-18.44,0-27.66-.04-9.06-6.71-10.76-12.79-8.26-1.95.8-3.53,2.56-5.21,3.96-.32.27-.47.9-.46,1.36.07,4.99.23,9.97.28,14.96.05,5.1-.05,10.2.05,15.3.03,1.45.49,2.9.76,4.39h-6.87c.28-1.1.77-2.16.77-3.22.04-9.79.09-19.57-.05-29.36-.12-8.13-6.65-9.62-12.2-7.62-2.29.82-4.11,2.32-5.63,4.17-.38.47-.63,1.19-.64,1.79-.04,10.04-.06,20.09,0,30.13,0,1.35.51,2.69.79,4.09h-6.75c.22-1.43.63-2.85.64-4.27.05-11.33.05-22.66,0-33.99,0-1.19-.53-2.38-.89-3.86Z"
+          />
+          <path
+            fill={color}
+            d="M.43.59c6.28.29,12.82.43,19.33.95,2.98.24,5.81,1.41,7.92,3.74,2.56,2.83,2.69,7.81.16,11.31-1.32,1.82-2.96,3.42-4.49,5.08-1.67,1.82-3.4,3.59-5.09,5.39-4.54,4.84-3.95,11.12-.82,16.07,3.22,5.11,7.93,8.72,12.8,12.12.8.56,1.63,1.08,2.44,1.96-1.23,0-2.48-.14-3.69.03-3.53.49-6.08-.87-8.56-3.37-6.68-6.72-8.82-18.73-4.26-26.98,1.88-3.41,4.42-6.46,6.29-9.88.99-1.82,1.56-4.05,1.68-6.13.22-3.82-2.2-6.74-6.04-7.97-3.69-1.18-7.49-1.14-11.46-1-.03.61-.09,1.15-.09,1.7,0,16.53-.02,33.06.02,49.59,0,1.27.4,2.54.63,3.89H.06c1.68-1.98,1.21-4.33,1.22-6.58.03-15.09.03-30.17-.02-45.26,0-1.41-.48-2.82-.82-4.65Z"
+          />
+          <path
+            fill={color}
+            d="M78.28,34.61c.25-8.28,4.17-15.39,12.67-19.57,8.51-4.18,20.13-.73,25.15,7.46,5.44,8.86,5.42,17.85-.27,26.54-3.42,5.21-8.52,8.09-14.82,8.55-8.77.64-15.32-2.97-19.81-10.44-1.88-3.12-2.79-6.85-2.92-12.54ZM83.99,35.59c-.04,3.64.55,7.17,1.94,10.54,1.69,4.12,4.17,7.64,8.36,9.46,5.82,2.53,11.95.41,15.74-5.17,5.66-8.35,5.62-21.6-.09-29.91-5.35-7.78-14.29-8.44-20.5-1.37-4.1,4.67-5.4,10.39-5.45,16.45Z"
+          />
+          <path
+            fill={color}
+            d="M273.05,55.72c-4.68-2.65-9.2-5.45-12.3-9.42-3.01-3.84-3.58-8.73.78-12.48,1.62-1.4,3.28-2.77,4.89-4.18,1.47-1.3,3.04-2.53,4.31-3.95,2.43-2.72,2.31-6.58-.16-8.78-2.03-1.81-4.75-2.72-7.61-2.9-6.26-.4-12.55-.51-18.58-.74.33,1.42.78,2.52.78,3.61.03,6.84.04,13.69.04,20.53-.32.65-.65,1.36-1,2.09-1.02,2.16-2.16,4.49-3.18,6.02-1.82,2.74-3.94,4.62-3.94,4.62l-.4.41-.02.02-.02.02c-1.2.98-2.2,1.51-3.58,2-8,2.83-16.23-.92-19.52-8.85-2.44-5.88-2.43-11.87-1.09-18.09h26.97c.45-4.48-1.96-9.02-5.62-10.71-8.09-3.74-18.01-.95-23.15,7.28-3.52,5.64-4.39,11.82-3.25,18.25,1.52,8.62,6.79,15.17,15.15,16.92,4.11.86,8.03.23,11.2-1.71.43-.26.85-.54,1.24-.85,0,0,0,0,.02,0,.57-.44,1.11-.93,1.61-1.47l.04-.04c.11-.11.3-.29.54-.52,0,0,0-.02.02-.02,1.07-1.03,3.21-3.17,4.77-5.67.75-1.2,1.5-2.49,2.22-3.81.34-.62.68-1.25,1-1.87,0,3.54,0,7.08,0,10.62,0,1.75.45,3.57-1.17,5.11h6.86c-.22-1.05-.6-2.03-.6-3.02-.04-12.84-.03-25.67-.02-38.51,0-.43.05-.85.09-1.32,3.81-.12,7.46-.14,11.02.78,3.7.95,6.02,3.21,5.81,6.19-.12,1.61-.66,3.35-1.61,4.76-1.79,2.66-4.23,5.02-6.04,7.68-4.38,6.4-2.32,15.74,4.1,20.95,2.38,1.93,4.84,3,8.22,2.61,1.16-.12,2.36-.02,3.55-.02-.78-.68-1.58-1.09-2.34-1.52ZM228.84,14.62c2.92.62,4.8,2.58,5.15,5.38.28,2.23-.64,3.68-2.76,4.32-.54.16-1.11.32-1.66.32-4.95.04-9.91.05-14.86.06h-1.81c2.02-7.38,8.8-11.59,15.94-10.08Z"
+          />
+          <path
+            fill={color}
+            d="M60.36,47.46c-.07.12-.15.23-.23.34.14-.16.28-.33.43-.51-.09.05-.16.1-.2.17Z"
+          />
+          <path
+            fill={color}
+            d="M75.14,4.67c3.42-.35,6.76.18,10.02,1.54.44-2.02-.28-3.45-1.72-4.56-2.8-2.17-7.08-2.14-10.32-.17-3.15,1.93-4.7,4.91-5.86,8.21-.82,2.32-2.28,3.87-4.88,4.11-.35.03-.69.2-1.03.3,0,.18.02.36.02.54,1.51.08,3.02.16,4.67.26v1.92c0,6.85,0,13.71,0,20.57-.33.66-.67,1.38-1.02,2.13-1.02,2.16-2.16,4.49-3.18,6.02-.43.64-.87,1.24-1.3,1.77-.15.18-.29.35-.43.51,0,.02-.02.02-.03.03-.83,1.26-1.85,2.28-2.98,3.1-4.01,2.9-9.46,3.1-13.74,1.65-4.34-1.47-7.17-4.55-8.91-8.62-2.55-5.92-2.67-12.03-1.14-18.31h15.6c3.41,0,6.82,0,10.22,0,.76,0,1.29-.07,1.27-1.06-.09-4.6-2.65-8.59-6.96-10.15-9.28-3.35-19.25.89-23.5,10.67-3.06,7.05-2.83,14.2.09,21.28,3.42,8.33,12.55,12.98,21.21,10.81,1.6-.4,3.21-1.27,4.63-2.4.09-.05.18-.09.26-.14,0,0,.02,0,.02,0,.67-.55,1.29-1.16,1.86-1.86,0,0,0-.02.02-.02,1.07-1.03,3.21-3.17,4.78-5.67.75-1.2,1.49-2.49,2.22-3.81.34-.63.68-1.27,1.02-1.89,0,3.49-.02,6.98-.03,10.47,0,1.76-.43,3.52-.66,5.27h6.61c-.21-1.57-.6-3.1-.61-4.62-.05-11.95-.02-23.89-.02-35.84v-1.82c4.01-.31,7.84.17,11.6-.65-.69-.37-1.37-.34-2.06-.37-3.23-.17-6.5-.1-9.67-.62-2.62-.43-3.47-3.03-2.07-5.27,1.37-2.2,3.56-3.04,5.99-3.28ZM37.81,17.82c3.87-3.43,8.35-4.48,13.29-2.66,2.59.95,3.94,3.17,3.8,5.73-.09,1.61-.92,2.69-2.4,3.25-.61.23-1.28.47-1.93.47-5.55.04-11.11.02-16.82.02.72-2.73,2.06-5.03,4.07-6.81Z"
+          />
+          <path
+            fill={color}
+            d="M143.72,13.5c-.29,1.53-.5,2.95-.88,4.33-.11.4-.71.88-1.14.94-5.91.92-9.05,4.78-10.51,10.15-.67,2.44-.89,5.05-.96,7.6-.14,5.4-.09,10.81,0,16.21.02,1.47.45,2.92.69,4.41h-6.64c.21-1.51.59-2.97.6-4.44.05-11.12.02-22.24.04-33.36,0-1.53.03-3-1.69-4.15,2.44-.66,4.54-1.24,6.87-1.87v14.75c3.15-10.9,7.78-14.58,13.63-14.57Z"
+          />
+        </g>
+        <g>
+          <path
+            fill={color}
+            d="M237.2,52.79c-.57.69-1.19,1.3-1.86,1.86,0,0,0,0-.02,0-.1.05-.21.11-.31.16.57-.44,1.11-.93,1.61-1.47l.04-.04c.11-.11.3-.29.54-.52Z"
+          />
+          <path
+            fill={color}
+            d="M60.1,47.84c-.83,1.26-1.85,2.28-2.98,3.1l.78-.79s.99-.88,2.2-2.31Z"
+          />
+          <path
+            fill={color}
+            d="M236.65,53.31l-.04.04c-.5.54-1.03,1.02-1.61,1.47,0,0-.02,0-.02,0-.4.31-.82.6-1.24.85.12-.05.24-.09.36-.14.42-.26.82-.56,1.2-.88,0,0,.02,0,.02,0,.68-.55,1.29-1.16,1.86-1.86-.25.23-.43.41-.54.52Z"
+          />
+        </g>
+      </g>
+    </g>
+  </svg>
+)
+
+export const Mente = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="39"
+    height="87"
+    viewBox="0 0 39 87"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      d="M14.0894 47.0197C14.3494 46.2197 14.5894 45.3797 14.8394 44.5797C18.2494 33.8797 22.8694 22.8897 23.7994 12.8397C24.1894 8.79967 23.9094 4.90967 22.8094 1.31967C22.5494 0.469672 23.4394 -0.260328 24.2494 0.0896716C29.0394 2.17967 32.5994 4.99967 34.8494 8.28967C40.7094 16.7997 38.9694 28.0797 35.4794 39.5597C35.2194 40.3597 34.9794 41.1997 34.7294 41.9997C31.3194 52.6997 26.6994 63.6897 25.7694 73.7397C25.3794 77.7797 25.6594 81.6697 26.7494 85.2597C27.0094 86.1097 26.1194 86.8397 25.3094 86.4897C20.5194 84.3997 16.9594 81.5797 14.7094 78.2897C8.84943 69.7797 10.5894 58.4997 14.0794 47.0197H14.0894Z"
+      fill={color}
+    />
+    <path
+      d="M5.7794 11.3696C7.6694 11.0596 9.4894 11.1496 11.2294 11.9996C14.0294 13.3796 15.5594 16.4096 14.9794 19.3696C14.3494 22.5596 11.7394 24.8496 8.5094 25.0396C8.1894 25.0596 7.8694 25.0396 7.5594 25.0396C4.5694 25.0296 2.1294 23.9596 0.729401 21.1696C-1.3506 17.0296 1.2294 12.1096 5.7894 11.3596L5.7794 11.3696Z"
+      fill={color}
+    />
+  </svg>
+)
+
+export const SolGreyPebble = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="35"
+    height="33"
+    viewBox="0 0 45 43"
+    fill={color}
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <g clip-path="url(#clip0_113_8)">
+      <path
+        d="M16.53 12.1C16 8.84997 15.2 4.80997 14.56 1.23997C14.4 0.329973 15.61 -0.140027 16.1 0.649973L21.59 9.46997C21.92 9.99997 22.69 9.99997 23.02 9.46997L28.68 0.399973C29.16 -0.380027 30.37 0.0699734 30.23 0.969973C29.75 4.07997 29.26 7.33997 28.69 11.08C28.59 11.72 29.23 12.23 29.84 11.99L39.64 8.04997C40.49 7.70997 41.2 8.76997 40.55 9.41997C38.33 11.65 36.01 14 33.34 16.69C32.88 17.15 33.1 17.94 33.73 18.1L43.99 20.69C44.87 20.91 44.83 22.18 43.93 22.34C40.82 22.88 37.57 23.44 33.86 24.09C33.22 24.2 32.94 24.96 33.36 25.46L40.02 33.48C40.6 34.18 39.81 35.18 38.99 34.77C36.32 33.42 33.65 32.07 30.98 30.72C30.42 30.44 29.76 30.84 29.76 31.47L29.7 42.18C29.7 43.06 28.51 43.34 28.11 42.56L23.6 33.74C23.31 33.18 22.53 33.12 22.16 33.63L15.91 42.37C15.4 43.09 14.27 42.64 14.39 41.76C14.87 38.39 15.37 34.88 15.95 30.85C16.04 30.23 15.44 29.73 14.84 29.93L4.40002 33.56C3.55002 33.86 2.89002 32.79 3.53002 32.16L10.57 25.26C11.04 24.8 10.82 24 10.19 23.84L0.630023 21.4C-0.249977 21.17 -0.199977 19.9 0.700023 19.75C3.80002 19.23 7.04002 18.69 10.78 18.07C11.45 17.96 11.72 17.13 11.24 16.65C8.78002 14.15 6.58002 11.92 4.47002 9.76997C3.81002 9.09997 4.58002 8.02997 5.43002 8.41997L15.35 13C15.96 13.28 16.64 12.77 16.53 12.1Z"
+        fill={color}
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_113_8">
+        <rect width="44.62" height="43" fill={color} />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export const SolMidnight = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="35"
+    height="33"
+    viewBox="0 0 45 43"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <g clip-path="url(#clip0_113_11)">
+      <path
+        d="M16.53 12.1C16 8.84997 15.2 4.80997 14.56 1.23997C14.4 0.329973 15.61 -0.140027 16.1 0.649973L21.59 9.46997C21.92 9.99997 22.69 9.99997 23.02 9.46997L28.68 0.399973C29.16 -0.380027 30.37 0.0699734 30.23 0.969973C29.75 4.07997 29.26 7.33997 28.69 11.08C28.59 11.72 29.23 12.23 29.84 11.99L39.64 8.04997C40.49 7.70997 41.2 8.76997 40.55 9.41997C38.33 11.65 36.01 14 33.34 16.69C32.88 17.15 33.1 17.94 33.73 18.1L43.99 20.69C44.87 20.91 44.83 22.18 43.93 22.34C40.82 22.88 37.57 23.44 33.86 24.09C33.22 24.2 32.94 24.96 33.36 25.46L40.02 33.48C40.6 34.18 39.81 35.18 38.99 34.77C36.32 33.42 33.65 32.07 30.98 30.72C30.42 30.44 29.76 30.84 29.76 31.47L29.7 42.18C29.7 43.06 28.51 43.34 28.11 42.56L23.6 33.74C23.31 33.18 22.53 33.12 22.16 33.63L15.91 42.37C15.4 43.09 14.27 42.64 14.39 41.76C14.87 38.39 15.37 34.88 15.95 30.85C16.04 30.23 15.44 29.73 14.84 29.93L4.39996 33.56C3.54996 33.86 2.88996 32.79 3.52996 32.16L10.57 25.26C11.04 24.8 10.82 24 10.19 23.84L0.629962 21.4C-0.250038 21.17 -0.200038 19.9 0.699962 19.75C3.79996 19.23 7.03996 18.69 10.78 18.07C11.45 17.96 11.72 17.13 11.24 16.65C8.77996 14.15 6.57996 11.92 4.46996 9.76997C3.80996 9.09997 4.57996 8.02997 5.42996 8.41997L15.35 13C15.96 13.28 16.64 12.77 16.53 12.1Z"
+        fill={color}
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_113_11">
+        <rect width="44.62" height="43" fill={color} />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export const SolPearl = ({ color = "currentColor", ...props }) => (
+  <svg
+    width="35"
+    height="33"
+    viewBox="0 0 45 43"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <g clip-path="url(#clip0_113_5)">
+      <path
+        d="M16.53 12.1C16 8.84997 15.2 4.80997 14.56 1.23997C14.4 0.329973 15.61 -0.140027 16.1 0.649973L21.59 9.46997C21.92 9.99997 22.69 9.99997 23.02 9.46997L28.68 0.399973C29.16 -0.380027 30.37 0.0699734 30.23 0.969973C29.75 4.07997 29.26 7.33997 28.69 11.08C28.59 11.72 29.23 12.23 29.84 11.99L39.64 8.04997C40.49 7.70997 41.2 8.76997 40.55 9.41997C38.33 11.65 36.01 14 33.34 16.69C32.88 17.15 33.1 17.94 33.73 18.1L43.99 20.69C44.87 20.91 44.83 22.18 43.93 22.34C40.82 22.88 37.57 23.44 33.86 24.09C33.22 24.2 32.94 24.96 33.36 25.46L40.02 33.48C40.6 34.18 39.81 35.18 38.99 34.77C36.32 33.42 33.65 32.07 30.98 30.72C30.42 30.44 29.76 30.84 29.76 31.47L29.7 42.18C29.7 43.06 28.51 43.34 28.11 42.56L23.6 33.74C23.31 33.18 22.53 33.12 22.16 33.63L15.91 42.37C15.4 43.09 14.27 42.64 14.39 41.76C14.87 38.39 15.37 34.88 15.95 30.85C16.04 30.23 15.44 29.73 14.84 29.93L4.40002 33.56C3.55002 33.86 2.89002 32.79 3.53002 32.16L10.57 25.26C11.04 24.8 10.82 24 10.19 23.84L0.630023 21.4C-0.249977 21.17 -0.199977 19.9 0.700023 19.75C3.80002 19.23 7.04002 18.69 10.78 18.07C11.45 17.96 11.72 17.13 11.24 16.65C8.78002 14.15 6.58002 11.92 4.47002 9.76997C3.81002 9.09997 4.58002 8.02997 5.43002 8.41997L15.35 13C15.96 13.28 16.64 12.77 16.53 12.1Z"
+        fill={color}
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_113_5">
+        <rect width="44.62" height="43" fill={color} />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export const WhatsappIcon = ({ color = "currentColor", ...props }) => (
+  <svg
+    id="Layer_1"
+    data-name="Layer 1"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 500 500"
+    {...props}
+  >
+    <path
+      fill={color}
+      d="M239.28,65.87c7.2,0,14.4,0,21.59,0,6.17.77,12.37,1.32,18.49,2.35,38.22,6.39,71.59,22.76,99.1,50.06,41.75,41.43,60.13,91.87,54.83,150.37-3.52,38.82-18.43,73.38-44.53,102.4-43.85,48.76-98.79,69.27-163.93,61.45-26.29-3.15-50.6-12.42-72.92-26.71-2.23-1.43-4.14-1.66-6.66-.84-21.17,6.88-42.39,13.63-63.59,20.41-1.07.34-2.17.57-3.57.93.47-1.47.79-2.49,1.13-3.5,6.74-20.08,13.42-40.19,20.28-60.22,1.07-3.11.92-5.42-1.04-8.22-15.33-21.92-25.23-46.15-29.77-72.5-1.32-7.64-1.95-15.4-2.89-23.11,0-6.48,0-12.96,0-19.43.21-1.03.54-2.06.61-3.1,2.07-28.57,10.24-55.31,25.12-79.74,29.22-47.94,71.47-77.39,126.78-87.75,6.92-1.3,13.96-1.93,20.94-2.87ZM309.43,353.95c5.97-1.14,11.2-1.51,16-3.15,26.07-8.91,34.92-21.5,35.68-45.55.12-3.82-1.29-6.37-4.75-8.02-13.62-6.48-27.15-13.16-40.82-19.54-7.36-3.43-11.5-2.07-16.38,4.57-3.4,4.64-6.77,9.3-10.32,13.83-5.85,7.45-9.67,8.38-18.35,4.83-30.09-12.3-52.52-33.33-69.41-60.71-3.88-6.29-3.68-8.4,1.27-14.02,3.4-3.87,6.97-7.61,10.18-11.64,2.94-3.69,5.68-7.74,3.57-12.81-6.4-15.36-12.85-30.7-19.52-45.93-.82-1.86-2.9-4.03-4.75-4.44-9.83-2.18-19.98-2.69-28.12,4.16-17.48,14.72-23.09,34.18-19.64,56.26,2.45,15.67,9.79,29.4,18.7,42.38,14.31,20.86,30.46,40.04,49.72,56.55,21.43,18.37,46.31,30.04,73.11,37.93,7.99,2.35,16.28,3.66,23.82,5.31Z"
+    />
+  </svg>
+)
