--- src/components/Header/NavBar/NavBar.tsx
@@ -1,11 +1,0 @@
-import DesktopNav from "./DesktopNav"
-import MobileNav from "./MobileNav"
-
-export default function NavBar() {
-  return (
-    <>
-      <DesktopNav />
-      <MobileNav />
-    </>
-  )
-}
