+++ src/components/routes/root/Header/NavBar/NavBar.tsx
@@ -1,0 +1,11 @@
+import DesktopNav from "./DesktopNav"
+import MobileNav from "./MobileNav"
+
+export default function NavBar() {
+  return (
+    <>
+      <DesktopNav />
+      <MobileNav />
+    </>
+  )
+}
