+++ src/lib/types.ts
@@ -1,0 +1,6 @@
+export type ServicioInfo = {
+  name: string
+  description?: string
+  intensity?: number | string
+  zone?: string
+}
