--- types.ts
@@ -1,6 +1,0 @@
-export type ServicioInfo = {
-  name: string
-  description?: string
-  intensity?: number | string
-  zone?: string
-}
